tonic: Add support for async event handlers
This adds support for async event handlers. The important feature
here is to make sure we handle promise and do not leave it
dangling.

Since tonic cannot do much with the exception here we rethrow
it outside of the context of the promise chain.

This will cause it to bubble to window.onerror.

This allows for integrating with common third party systems like
Sentry which listens to window.onerror and sends all the errors
that happen in a website or web application to the backend
system of sentry itself.

Without handling this promise exception it becomes an unhandled
rejection which is more challenging to debug.