tonic: Allow component ref in props
This change sets a new top level field on the Tonic class
which is the `isTonicComponent` boolean for detecting whether
an instance of `Element` is a tonic component or a normal DOM
element.

We also allow for storing a reference to a tonic component if
it's passed as a property in `this.html()` . Similarly to how
we store a reference to non-trivial objects we can now store
a reference to a parent component so that it can be accessed from
`this.props`.