refs: Store ref by id instead of elem
When it comes to refs and persisting the state of the tonic
components in `Tonic._states` we should be doing so based on
`id` and not based on DOM element references.

This fixes a bug where we would lose the state of a tonic
component if it has been removed from the DOM.

I found this bug because there are two strategies for conditional
rendering of HTML elements

One of them is two just always render both HTML elements and make
one of the display: none.

The other is to have the template have a condition where we only
render one element or the other element, this causes them to be
swapped out. Since they are swapped out, one of the two child
tonic components will have it's disconnectedCallback called and
will be removed from the `_refs` array. This is not the behavior
that we actually want.

Instead when we are checking which element we want to save when
it comes to fetching their `getState()` we should just check
them by id instead.