index: Debounce reRender
When `reRender()` is scheduled it's possible for another reRender
to be scheduled in quick succession.

An example of this is handling both `click` and `change` event
at effectively the same time.

By typing in an input and then clicking a button it will fire both
a blur event on the input ( and change ) and a click event on the
button. We might have both a click & change handler in a component
which could cause two reRenders in effectively the same tick.