src: Fix reRender + updated event order bug
There is a subtle bug related to `reRender()` and `updated`.

We attempt to invoke the `updated` lifecycle hook on a component
if the `reRender()` has finished and the props are updated.

However, we null out the `pendingReRender` too late, since the
`updated` method can dispatch a DOM event ( e.g. `changed` )
which can trigger a reRender in an event listener.

We should set our internal state correctly before calling the
lifecycle `updated` method.

This fixes a bug in `components/checkbox`.