src: Do not cloneNode()
Doing so can lead to many unnecessary renders.
Especially with components like `tonic-split` and `tonic-tabs`
where they use `${this.elements}` or they detect for left & right.

This is because `render()` get's called multiple times, especially
if the render is `async` since the `cloneNode()` will not have
copied the children that are asynchronously rendered.

By having render called more times then expected, the component
will `async render()`; aka fetch and render multiple times (2 or 4)
and that leads to lots of unnecessary delays in interactivity.