src: Check the coupling between state & id.
Internally Tonic will only persist the `state` of a TonicComponent
if it has a `this.id` field.

There are also numerous components that set the `id` on a HTML
tagName with something like `input-${this.id}` ; this leads to
`input-undefined` if the user forgot the mandatory id.

This change makes accessing `this.state` and `this.id` throw
if there is no id.