src: Remove innerHTML assignment
We used to persist the previous innerHTML and re-assign it.

This breaks moving dom elements around like for example
in the tab component we add and remove elements from the DOM
when it comes to marking them visible or not.

I don't know why we assign innerHTML here when a few lines
lower we call _set() which will also assign innerHTML
with the result of calling render()